+++ misc/metrics/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.11.0 [unreleased]
+
+- Update to `libp2p-dcutr` `v0.8.0`.
+
+- Update to `libp2p-identify` `v0.41.0`.
+
+- Update to `libp2p-relay` `v0.14.0`.
+
 # 0.10.0
 
 - Update to `libp2p-swarm` `v0.40.0`.