+++ protocols/dcutr/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.8.0 [unreleased]
+
+- Update to `prost-codec` `v0.3.0`.
+
 # 0.7.0
 
 - Update to `libp2p-core` `v0.37.0`.